Implemented heat diffuse algorithm (TM) for braille chart!
I don't care that this is apparently already a thing for audio peak
detection, I thought of it independently, I get to name it!..
independently!

Seriously think about it, *heat*, what does temperature do? Or rather
what *is* temperature? What happens when two systems interact and there
is heat exchange? What hapens when something gets heated up, like the
barrel of a gun, and then is left in the air, cooling down?

Then think of the upper (hot) and lower (cold) bounds of a chart, and it
just makes so much _seeeeeeeeeeeeeeeense_

Heat Bounded Diffusal..

Bounded Heat Diffusal...

Adaptive Heat Diffusion...?

Diffuse Heat Transform!

..no

IDFK OK it works and I thought of it!

.. Activity Temperature Diffusal?